add validation for index

+++ 0x14-bit_manipulation/3-set_bit.c
@@ -12,8 +12,9 @@ int set_bit(unsigned long int *n, unsigned int index)
 {
 	unsigned long int num = 1;
 
-	if (index > sizeof(unsigned int) * 8)
+	if (index > sizeof(unsigned long int) * 8)
 		return (-1);
+
 	num = num << index;
 	*n = *n | num;
 	return (1);
